Move common functions into separate file
